# src/main/swagger/ShrimpFarm.yml
# this is an example of the Uber API
# as a demonstration of an API spec in YAML
swagger: '2.0'
info:
  title: API
  description: A shitty service that provides nothing.
  version: "1.0.0"
# the domain of the service
host: api.chaty.com
# array of all schemes that your API supports
schemes:
  - https
# will be prefixed to all paths
basePath: /v1
produces:
  - application/json
paths:
  /seasonSummary/{year}/{season}/:
    get:
      summary: Season Summary
      description: |
        This endpoint retuns a summary of information for the farming season selected.
      parameters:
        - name: year
          in: path
          description: Season Year
          required: true
          type: number
        - name: season
          in: path
          description: Season - One/two/three
          required: true
          type: number
      tags:
        - SeasonSummary
      responses:
        200:
          description: Summary of Shrimp Farm Season
          schema:
            type: array
            items:
              $ref: '#/definitions/SeasonProgress'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  /feedEntry:
    post:
      summary: Post Feed Entry
      description: |
        This endpoint adds a feed entry for the day/year/season.
      parameters:
        - name: FeedEntry
          in: body
          description: FeedEntry Object to be added.
          schema:
            $ref: '#/definitions/FeedEntry'
      tags:
        - PondActivity
      responses:
        200:
          description: Feed Entry Added Sucessfully.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  /stockEntry:
    post:
      summary: Post Stocking Entry
      description: |
        This endpoint adds a stock entry for the day/year/season.
      parameters:
        - name: StockEntry
          in: body
          description: StockEntry Object to be added.
          schema:
            $ref: '#/definitions/StockEntry'
      tags:
        - PondActivity
      responses:
        200:
          description: Stock Entry Added Sucessfully.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error' 
            
  /supplimentEntry:
    post:
      summary: Post Suppliment Entry
      description: |
        This endpoint adds a suppliment entry for the day/year/season.
      parameters:
        - name: SupplimentEntry
          in: body
          description: SupplimentEntry Object to be added.
          schema:
            $ref: '#/definitions/SupplimentEntry'
      tags:
        - PondActivity
      responses:
        200:
          description: Suppliment Entry Added Sucessfully.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error' 
            
  /expenseEntry:
    post:
      summary: Post Expense Entry
      description: |
        This endpoint adds a Expense entry for the day/year/season.
      parameters:
        - name: ExpenseEntry
          in: body
          description: ExpenseEntry Object to be added.
          schema:
            $ref: '#/definitions/ExpenseEntry'
      tags:
        - Monetory
      responses:
        200:
          description: Expense Entry Added Sucessfully.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
            
  /harvestEntry:
    post:
      summary: Post Sale Entry
      description: |
        This endpoint adds a Sale entry for the Pond/year/season.
      parameters:
        - name: SaleEntry
          in: body
          description: SaleEntry Object to be added.
          schema:
            $ref: '#/definitions/HarvestEntry'
      tags:
        - Monetory
      responses:
        200:
          description: Sale Entry Added Sucessfully.
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  
definitions:

  Pond:
    type: object
    properties:
      _id:
        type: string
        description: Monfo DB ID
      number:
        type: string
        description: Unique identifier representing a pond.
      season:
        type: string
        description: Feeding season format 2017-02.
      site:
        type: string
        description: Site where Pond is.
      size:
        type: string
        description: size of pond in acres.
      nickname:
        type: string
        description: pond nickname.
      stock:
        type: string
        description: Current Stock In Pond.
      feed:
        type: string
        description: Current Feed Consumption
      status:
        type: string
        description: Current Pond Status Good/Bad/Shit/BatShit.
        
  FeedEntry:
    type: object
    properties:
      _id:
        type: string
        description: Monfo DB ID
      pond:
        type: string
        description: Pond number
      date:
        type: string
        description: mm/dd/yyyy format date String.
      season:
        type: string
        description: Feeding season format 2017-02.
      site:
        type: string
        description: Site where Pond is.
      type:
        type: string
        description: Feed type.
      amount:
        type: number
        description: feed in kilograms.
      check:
        type: string
        description: complete/partial/bad.
      time:
        type: string
        description: morning/aft/evening/night.
  
  SupplimentEntry:
    type: object
    properties:
      _id:
        type: string
        description: Monfo DB ID
      pond:
        type: string
        description: Pond number
      date:
        type: string
        description: mm/dd/yyyy format date String.
      season:
        type: string
        description: Feeding season 2017-02.
      site:
        type: string
        description: Site where Pond is.
      type:
        type: string
        description: Supplement type.
      planned:
        type: boolean
        description: Is Planned
      amount:
        type: number
        description: feed in kilograms.
      time:
        type: string
        description: morning/aft/evening/night.
        
  StockEntry:
    type: object
    properties:
      _id:
        type: string
        description: Monfo DB ID
      pond:
        type: string
        description: Pond number
      date:
        type: string
        description: mm/dd/yyyy format date String.
      season:
        type: string
        description: Feeding season 2017-02.
      site:
        type: string
        description: Site where Pond is.
      salinity:
        type: number
        description: Water Salinity.
      amount:
        type: number
        description: number of seed.
      hatchery:
        type: string
        description: Hatchery where seed is bought.
      plSize:
        type: integer
        description: Post Larva Size. 
        
  HarvestEntry:
    type: object
    properties:
      _id:
        type: string
        description: Monfo DB ID
      pondNumber:
        type: string
        description: Pond number
      date:
        type: string
        description: mm/dd/yyyy format date String.
      season:
        type: string
        description: Feeding season 2017-02.
      buyer:
        type: string
        description: Feed type.
      planned:
        type: boolean
        description: Is Planned Harvest.
      tonnage:
        type: number
        description: Tonnage in kilograms.
      countPrice:
        type: number
        description: Price of Harvested count Size.
      amount:
        type: number
        description: Total Harvest Amount
      comments:
        type: string
        description: Comments
        
  ExpenseEntry:
    type: object
    properties:
      _id:
        type: string
        description: Monfo DB ID
      category:
        type: string
        description: expense category name.
      quantity:
        type: string
        description: expense item quantity.
      isPlanned:
        type: string
        description: Yes/No
      paymentType:
        type: string
        description: Yes/No
      price:
        type: integer
        description: expense price in rupees.
      date:
        type: string
        description: mm/dd/yyyy expense date.
      comments:
        type: string
        description: paid by/paid to/cash/bank/cheque
        
        
  FarmProgress:
    type: object
    properties:
      pondNumber:
        type: integer
        format: int32
        description: Pond Number.
      daysOfCultivation:
        type: integer
        format: int32
        description: Days of cultivation.
      count:
        type: integer
        format: int32
        description: Shrimp Count.
      feedConsumption:
        type: integer
        description: Current Feed Consumption.
      expWeight:
        type: integer
        format: int32
        description: Expected shrimp weight.
        
  SeasonProgress:
    type: object
    properties:
      farmProgressList:
        type: array
        description: Pond Number.
        items:
          $ref: '#/definitions/FarmProgress'
      daysOfCultivation:
        type: integer
        format: int32
        description: Days of cultivation.
      count:
        type: integer
        format: int32
        description: Shrimp Count.
      feedConsumption:
        type: integer
        description: Current Feed Consumption.
      expWeight:
        type: integer
        format: int32
        description: Expected shrimp weight.
     
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string